Modify directory structure
