+++ script.rb
@@ -1,6 +1,6 @@
-require './sophian'
-require './lib/exporter'
-require './lib/grade_exporter'
+require './lib/sophian'
+require './lib/exporter/grade_exporter'
+require 'pry'
 
 def crawl_syllabus(sophian)
   sophian.crawl_syllabus_from_loyola
@@ -27,8 +27,7 @@ def export_crawl_result(crawl_results, exporter_instance)
   puts '何か文字を入力するとエクスポートを開始します。'
   exporter_instance.execute(crawl_results) if $stdin.gets.chomp
 end
-
-
+binding.pry
 sophian = Sophian.new(id: ENV['LOYOLA_LOGIN_ID'], password: ENV['LOYOLA_LOGIN_PASSWORD'])
 
 puts <<EOS
